+++ packages/next-templates/src/app/wmebv/DigiD_inloggen/page.tsx
@@ -8,7 +8,7 @@ import {
   UtrechtHeading1,
   UtrechtButtonGroup,
 } from '@utrecht/web-component-library-react';
-import { ExampleHeaderWmebv } from '@/components/ExampleHeader/wmebv/ExampleHeaderWmebv';
+import { ExampleHeaderWmebv } from '@/components/wmebv/Header/ExampleHeaderWmebv';
 import { ExampleNavigation } from '@/components/ExampleNavigation/ExampleNavigation';
 import '@/app/styling/css/wmebv.css';
