+++ packages/next-templates/src/app/wmebv/ingelogd/stap3/page.tsx
@@ -18,7 +18,7 @@ import {
   UtrechtParagraph,
 } from '@utrecht/web-component-library-react';
 import { LinkButton, PreHeading, HeadingGroup } from '@utrecht/component-library-react';
-import { ExampleHeaderFunnelWmebv } from '@/components/ExampleHeader/wmebv/ExampleHeaderFunnelWmebv';
+import { ExampleHeaderFunnelWmebv } from '@/components/wmebv/Header/ExampleHeaderFunnelWmebv';
 import { ExampleFooter } from '@/components/ExampleFooter/ExampleFooter';
 import ArrowLeft from '@/app/styling/assets/arrow-left-icon.svg';
 import Pencil from '../../../styling/assets/pencil-icon.svg';
