+++ packages/next-templates/src/app/wmebv/ingelogd/stap4/page.tsx
@@ -11,7 +11,7 @@ import {
   UtrechtSpotlightSection,
   UtrechtButtonGroup,
 } from '@utrecht/web-component-library-react';
-import { ExampleHeaderWmebv } from '@/components/ExampleHeader/wmebv/ExampleHeaderWmebv';
+import { ExampleHeaderWmebv } from '@/components/wmebv/Header/ExampleHeaderWmebv';
 import { ExampleFooter } from '@/components/ExampleFooter/ExampleFooter';
 import { ExampleNavigation } from '@/components/ExampleNavigation/ExampleNavigation';
 import { UnorderedList, UnorderedListItem } from '@utrecht/component-library-react';
